⚡ Bolt: Replace strcmp with inline helper for directory iteration
💡 What: Replaced `strcmp(name, ".") == 0 || strcmp(name, "..") == 0` with an explicit `is_dot_or_dotdot` manual byte-comparison inline helper in hot directory iteration loops. Also fixed several Xcode build warnings causing CI failures.
🎯 Why: `strcmp` calls have function overhead. Replacing them with direct character comparisons converts an O(N) function call into a constant-time O(1) array access, which is highly beneficial in hot loops like VFS directory iterations. Xcode build warnings were causing CI failures due to -Werror.
📊 Impact: Expected to reduce directory iteration time and function overhead. CI Xcode build pass.
🔬 Measurement: Use `perf` or `time` to measure directory reading times before and after this optimization.

+++ app/AppDelegate.m
@@ -3239,7 +3239,7 @@ static TerminalViewController *CreateTerminalViewController(void) {
                 [self scheduleDnsRefresh:@"dnsnotify"];
             });
             if (err == NOTIFY_STATUS_OK) {
-                self.dnsNotifyToken = (int) token;
+                self.dnsNotifyToken = token;
                 self.dnsNotifyRegistered = YES;
             }
         }
